requirements.txt
allure-pytest==2.9.43
allure-python-commons==2.9.43
arrow==1.1.1
atomicwrites==1.4.0
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
colorama==0.4.4
dateparser==1.0.0
docopt==0.6.2
idna==2.10
iniconfig==1.1.1
loguru==0.5.3
moment==0.12.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
python-dateutil==2.8.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
selenium==3.141.0
selenium-page-factory==2.3
six==1.16.0
times==0.7
toml==0.10.2
tzlocal==2.1
urllib3==1.26.6
win32-setctime==1.0.3
yarg==0.1.9
